# docs/swagger.yaml
basePath: /api/v1
definitions:
  ErrorResponse:
    properties:
      msg:
        example: error message text
        type: string
      status:
        example: error
        type: string
    type: object
  SuccessResponse:
    properties:
      data:
        type: object
      status:
        example: success
        type: string
    type: object
  models.Agent:
    properties:
      connected_date:
        type: string
      created_date:
        type: string
      description:
        type: string
      hash:
        type: string
      id:
        type: integer
      info:
        $ref: '#/definitions/models.AgentInfo'
      ip:
        type: string
      status:
        type: string
    required:
    - description
    - hash
    - info
    - ip
    - status
    type: object
  models.AgentInfo:
    properties:
      os:
        $ref: '#/definitions/models.AgentOS'
      user:
        $ref: '#/definitions/models.AgentUser'
    required:
    - os
    - user
    type: object
  models.AgentOS:
    properties:
      arch:
        type: string
      name:
        type: string
      type:
        type: string
    required:
    - arch
    - name
    - type
    type: object
  models.AgentUser:
    properties:
      group:
        type: string
      name:
        type: string
    required:
    - group
    - name
    type: object
...
  models.Group:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
...
  models.Password:
    properties:
      confirm_password:
        type: string
      current_password:
        type: string
      password:
        type: string
    required:
    - current_password
    - password
    type: object
...
  models.UserGroup:
    properties:
      group:
        $ref: '#/definitions/models.Group'
      group_id:
        type: integer
      id:
        type: integer
      mail:
        type: string
      name:
        type: string
      status:
        type: string
      tenant_id:
        type: integer
    required:
    - mail
    - status
    type: object
  private.agent:
    properties:
      agent:
        $ref: '#/definitions/models.Agent'
      details:
        $ref: '#/definitions/private.agentDetails'
    type: object
  private.agentDetails:
    properties:
      active_modules:
        type: integer
      events_per_last_day:
        type: integer
      hash:
        type: string
    type: object
  private.agentInfo:
    properties:
      arch:
        default: amd64
        enum:
        - "386"
        - amd64
        type: string
      name:
        type: string
      os:
        default: linux
        enum:
        - windows
        - linux
        - darwin
        type: string
    required:
    - arch
    - name
    - os
    type: object
...
  private.agents:
    properties:
      agents:
        items:
          $ref: '#/definitions/models.Agent'
        type: array
      details:
        items:
          $ref: '#/definitions/private.agentDetails'
        type: array
    type: object
...
info:
  contact:
    email: admin@vxcontrol.com
    name: Dmitry Nagibin
    url: https://vxcontrol.com
  description: Swagger API for VXControl backend product.
  license:
    name: Proprietary License
    url: https://github.com/vxcontrol/vxui/src/master/LICENSE
  termsOfService: http://swagger.io/terms/
  title: VXUI Swagger API
  version: "1.0"
paths:
  /agents/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: agents list received successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/private.agents'
              type: object
        "403":
          description: getting agents not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: agents not found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Retrieve agents list
      tags:
      - Agents
    put:
      consumes:
      - application/json
      parameters:
      - description: agent info to create one
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/private.agentInfo'
      produces:
      - application/json
      responses:
        "201":
          description: agent created successful
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: invalid agent info
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: creating agent not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create new agent in service
      tags:
      - Agents
  /agents/{hash}:
    delete:
      parameters:
      - description: agent hash in hex format (md5)
        in: path
        maxLength: 32
        minLength: 32
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: agent deleted successful
          schema:
            $ref: '#/definitions/SuccessResponse'
        "403":
          description: deleting agent not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: agent not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Delete agent from instance DB
      tags:
      - Agents
    get:
      parameters:
      - description: agent hash in hex format (md5)
        in: path
        maxLength: 32
        minLength: 32
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: agent info received successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/private.agent'
              type: object
        "403":
          description: getting agent info not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: agent not found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Retrieve agent info by agent hash
      tags:
      - Agents
    post:
      consumes:
      - application/json
      parameters:
      - description: agent hash in hex format (md5)
        in: path
        maxLength: 32
        minLength: 32
        name: hash
        required: true
        type: string
      - description: agent info as JSON data
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.Agent'
      produces:
      - application/json
      responses:
        "200":
          description: agent info updated successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/private.agent'
              type: object
        "400":
          description: invalid agent info
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: updating agent info not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: agent not found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update agent info by agent hash
      tags:
      - Agents
...
  /users/current:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: user info received successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserGroup'
              type: object
        "403":
          description: getting user not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Retrieve current user information
      tags:
      - Users
  /users/current/password:
    post:
      consumes:
      - application/json
      parameters:
      - description: container to validate and update account password
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.Password'
      produces:
      - application/json
      responses:
        "200":
          description: account password updated successful
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: account password form data invalid
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: updating account password not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on updating account password
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update password for current user (account)
      tags:
      - Users
swagger: "2.0"